fix: improved use of mvvm pattern in containerconsole
Much of the logic was taken from the temporary experimental project I created that didn't follow MVVM guidelines (to understand docker.dotnet better), so mistranslations occurred when moving it here. I noticed that a lot of logic was defined in the view instead of the viewmodel, so this is an attempt to fix that.