fix: alternative solution of handling auto-scroll
The previous implementation scrolled to the last line posted in the terminal. However, since the lines are added to a queue (for performance reasons) and processed as the UI thread has time, it is not guaranteed that the last line gets posted before auto-scrolling is done. This resulted in some unexpected behavior as it wasn't guaranteed to auto-scroll to the last line.

This new approach uses an event that is invoked when important lines are posted to the UI (handle accordingly, ex scroll the the line).