fix: improved the performance by using events
The preivous implementaiton using a "Output" variable in the viewmodel was realllly slow since the textbox was fully updated for each line added (hot-reload). The new approach uses a event that is invoked with a buffer of new lines to add. The view is subscribed to the event and will only add the newlines to the textbox.